finished prove05

+++ controllers/admin.js
@@ -1,26 +1,36 @@
 const Product = require('../models/product');
 
 exports.getAddProduct = (req, res) => {
+  if (!req.session.isLoggedIn){
+    return res.redirect('/auth')
+  }
   Product.find()
   .then(products => {
     res.render('pages/admin', {
       pageTitle: 'Add Product',
       path: '/admin',
       prods: products,
-      test: "testtest"
+      isAuth: req.session.isLoggedIn
     });
   })
   .catch(err => {console.log(err);})
   };
   //https://drive.google.com/uc?id=
 
+
+
+
+
   exports.postAddProduct =(req, res) => {
+    if (!req.session.isLoggedIn){
+      return res.redirect('/auth')
+    }
     const title = req.body.title;
     const description = req.body.description;
     const price = req.body.price;
     const tag = req.body.tag;
     const imgUrl = req.body.imgUrl;
-    console.log(req.user)
+    //console.log(req.session.user._id);
     const product = new Product({title: title, description: description, price: price, tag: tag, imgUrl: imgUrl, userId: req.user._id});
     product.save()
       .then(result => {
@@ -32,6 +42,9 @@ exports.getAddProduct = (req, res) => {
   };
 
   exports.getEdit = (req, res) =>{
+    if (!req.session.isLoggedIn){
+      return res.redirect('/auth')
+    }
     let title = "";
     let price = 0;
     let description = "";
@@ -55,13 +68,17 @@ exports.getAddProduct = (req, res) => {
                 itemD: description,
                 itemI: imgUrl,
                 ID: id,
-                TAG: tag
+                TAG: tag,
+                isAuth: req.session.isLoggedIn
             });
         })
         .catch(err => console.log(err));
   };
 
   exports.postEdit = (req, res) => {
+    if (!req.session.isLoggedIn){
+      return res.redirect('/auth')
+    }
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
@@ -85,6 +102,9 @@ exports.getAddProduct = (req, res) => {
   };
   
   exports.postDeleteProduct = (req, res) => {
+    if (!req.session.isLoggedIn){
+      return res.redirect('/auth')
+    }
     const prodId = req.body.deleteId
     Product.findByIdAndRemove(prodId)
       .then(() => {
